Refactor: split auth, onboarding and intro. Remove superflous intro pages
