+++ app/lib/common/dialogs/logout_confirmation.dart
@@ -1,7 +1,7 @@
 import 'package:acter/common/models/keys.dart';
 import 'package:acter/common/toolkit/buttons/danger_action_button.dart';
 
-import 'package:acter/features/onboarding/providers/onboarding_providers.dart';
+import 'package:acter/features/auth/providers/auth_providers.dart';
 import 'package:atlas_icons/atlas_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
