+++ app/lib/common/utils/routes.dart
@@ -5,13 +5,16 @@ enum Routes {
   actionAddEvent('/actions/addEvent'),
   actionCreateSuperInvite('/actions/createSuperInvite'),
 
-  // --- Auth
+  // pre
   intro('/intro'),
-  start('/start'),
   introProfile('/introProfile'),
+
+  // --- Auth
   authLogin('/login'),
   forgotPassword('/forgotPassword'),
   authRegister('/register'),
+
+  // -- onboarding
   saveUsername('/saveUsername'),
   linkEmail('/linkEmail'),
   uploadAvatar('/uploadAvatar'),
