+++ app/lib/features/auth/pages/login_page.dart
@@ -5,8 +5,8 @@ import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
 import 'package:acter/common/utils/constants.dart';
 import 'package:acter/common/utils/routes.dart';
 import 'package:acter/common/widgets/no_internet.dart';
-import 'package:acter/features/onboarding/providers/onboarding_providers.dart';
-import 'package:acter/features/onboarding/widgets/logo_widget.dart';
+import 'package:acter/features/auth/providers/auth_providers.dart';
+import 'package:acter/features/auth/widgets/logo_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
