+++ app/lib/features/auth/providers/auth_providers.dart
@@ -1,4 +1,4 @@
-import 'package:acter/features/onboarding/providers/notifiers/auth_notifier.dart';
+import 'package:acter/features/auth/providers/notifiers/auth_notifier.dart';
 import 'package:riverpod/riverpod.dart';
 
 final authStateProvider = StateNotifierProvider<AuthStateNotifier, bool>(
