+++ app/lib/features/auth/widgets/logo_widget.dart
