+++ app/lib/l10n/app_de.arb
@@ -1816,12 +1816,14 @@
   "@chatSettingsReadReceipts": {},
   "chatSettingsReadReceiptsExplainer": "(bald) Lasse die anderen wissen, wenn du eine Nachricht gelesen hast",
   "@chatSettingsReadReceiptsExplainer": {},
-  "introPageDescription1": "Acter ist mehr als eine App\nEs ist",
-  "@introPageDescription1": {},
-  "introPageDescription2": " - Community von ChangeMarkern.",
-  "@introPageDescription2": {},
-  "introPageDescription3": "Verbinde dich mit anderen Aktivisti, teile Erkenntnisse und kollaboriere an bedeutsamen Projekten.",
-  "@introPageDescription3": {},
+  "introPageDescriptionPre": "Acter ist mehr als eine App\nEs ist",
+  "@introPageDescriptionPre": {},
+  "introPageDescriptionHl": " - Community von ChangeMarkern.",
+  "@introPageDescriptionHl": {},
+  "introPageDescriptionPost": "",
+  "@introPageDescriptionPost": {},
+  "introPageDescription2ndLine": "Verbinde dich mit anderen Aktivisti, teile Erkenntnisse und kollaboriere an bedeutsamen Projekten.",
+  "@introPageDescription2ndLine": {},
   "logOutConformationDescription1": "Vorsicht ",
   "@logOutConformationDescription1": {},
   "logOutConformationDescription2": "Sich abmelden entfernt die lokalen Daten, inklusive der Verschlüsselungsdaten. Wenn dies deine letzte eingeloggte Sitzung ist, kann es sein, dass du dadurch den Zugang zu allen bisherigen Daten verlierst.",
