+++ app/lib/l10n/app_en.arb
@@ -1869,12 +1869,14 @@
   "@createOrJoinSpaceDescription": {},
   "spaceShortDescription": "a space, to start organizing and collaborating!",
   "@spaceShortDescription": {},
-  "introPageDescription1": "Acter is more than just an app.\nIt’s",
-  "@introPageDescription1": {},
-  "introPageDescription2": " community of change makers.",
-  "@introPageDescription2": {},
-  "introPageDescription3": "Connect with fellow activists, share insights, and collaborate on creating meaningful change.",
-  "@introPageDescription3": {},
+  "introPageDescriptionPre": "Acter is more than just an app.\nIt’s",
+  "@introPageDescriptionPre": {},
+  "introPageDescriptionHl": " community of change makers.",
+  "@introPageDescriptionHl": {},
+  "introPageDescriptionPost": "",
+  "@introPageDescriptionPost": {},
+  "introPageDescription2ndLine": "Connect with fellow activists, share insights, and collaborate on creating meaningful change.",
+  "@introPageDescription2ndLine": {},
   "logOutConformationDescription1": "Attention: ",
   "@logOutConformationDescription1": {},
   "logOutConformationDescription2": "Logging out removes the local data, including encryption keys. If this is your last signed-in device you might no be able to decrypt any previous content.",
