+++ app/lib/l10n/app_pl.arb
@@ -1576,8 +1576,8 @@
   "@changeThePowerFromTo": {},
   "spaceShortDescription": "przestrzeń, aby zacząć organizować i współpracować!",
   "@spaceShortDescription": {},
-  "introPageDescription1": "Acter to coś więcej niż tylko aplikacja",
-  "@introPageDescription1": {},
+  "introPageDescriptionPre": "Acter to coś więcej niż tylko aplikacja",
+  "@introPageDescriptionPre": {},
   "logOutConformationDescription2": "Wylogowanie usuwa lokalne dane, w tym klucze szyfrowania. Jeśli jest to twoje ostatnie zalogowane urządzenie, możesz nie być w stanie odszyfrować poprzedniej zawartości.",
   "@logOutConformationDescription2": {},
   "activitiesDescription": "Wszystkie ważne rzeczy wymagające uwagi można znaleźć tutaj",
@@ -1886,10 +1886,12 @@
   "@unlinkRoom": {},
   "createOrJoinSpaceDescription": "Stwórz przestrzeń lub dołącz do niej, aby zacząć organizować i współpracować!",
   "@createOrJoinSpaceDescription": {},
-  "introPageDescription2": " społeczność twórców zmian.",
-  "@introPageDescription2": {},
-  "introPageDescription3": "Nawiązuj kontakty z innymi aktywistami, dziel się spostrzeżeniami i współpracuj przy znaczących projektach.",
-  "@introPageDescription3": {},
+  "introPageDescriptionHl": " społeczność twórców zmian.",
+  "@introPageDescriptionHl": {},
+  "introPageDescriptionPost": "",
+  "@introPageDescriptionPost": {},
+  "introPageDescription2ndLine": "Nawiązuj kontakty z innymi aktywistami, dziel się spostrzeżeniami i współpracuj przy znaczących projektach.",
+  "@introPageDescription2ndLine": {},
   "logOutConformationDescription1": "Uwaga: ",
   "@logOutConformationDescription1": {},
   "logOutConformationDescription3": " Czy na pewno chcesz się wylogować?",
