+++ app/test/features/auth/pages/login_page_test.dart
@@ -1,5 +1,5 @@
 import 'package:acter/common/utils/constants.dart';
-import 'package:acter/features/onboarding/pages/login_page.dart';
+import 'package:acter/features/auth/pages/login_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
